+++ packages/forest_admin_datasource_rpc/lib/forest_admin_datasource_rpc.rb
@@ -19,11 +19,16 @@ module ForestAdminDatasourceRpc
   #   - Valid range: 1-3600 seconds
   #   - Priority: options[:schema_polling_interval] > ENV['SCHEMA_POLLING_INTERVAL_SEC'] > default
   #   - Example: SCHEMA_POLLING_INTERVAL_SEC=30 for development (30 seconds)
+  # @option options [Hash] :introspection Pre-defined schema introspection for resilient deployment
+  #   - When provided, allows the datasource to start even if the RPC slave is unreachable
+  #   - The introspection will be used as fallback when the slave connection fails
+  #   - Schema polling will still be enabled to pick up changes when the slave becomes available
   #
   # @return [ForestAdminDatasourceRpc::Datasource] The configured datasource with schema polling
   def self.build(options)
     uri = options[:uri]
     auth_secret = options[:auth_secret] || ForestAdminAgent::Facades::Container.cache(:auth_secret)
+    provided_introspection = options[:introspection]
 
     # Create schema polling client with configurable polling interval
     # Priority: options[:schema_polling_interval] > ENV['SCHEMA_POLLING_INTERVAL_SEC'] > default (600)
@@ -39,7 +44,7 @@ module ForestAdminDatasourceRpc
       polling_interval: polling_interval
     }
 
-    schema_polling = Utils::SchemaPollingClient.new(uri, auth_secret, polling_options) do
+    schema_polling = Utils::SchemaPollingClient.new(uri, auth_secret, polling_options, provided_introspection) do
       # Callback when schema change is detected
       logger = ForestAdminAgent::Facades::Container.logger
       logger.log('Info', '[RPCDatasource] Schema change detected, reloading agent...')
@@ -53,48 +58,22 @@ module ForestAdminDatasourceRpc
     # Get the schema from the polling client
     schema = schema_polling.current_schema
 
+    # Use provided introspection as fallback when slave is unreachable
+    if schema.nil? && provided_introspection
+      ForestAdminAgent::Facades::Container.logger.log(
+        'Warn',
+        "RPC agent at #{uri} is unreachable, using provided introspection for resilient deployment."
+      )
+      options.delete(:introspection)
+      schema = provided_introspection
+    end
+
     if schema.nil?
       # return empty datasource for not breaking stack
       ForestAdminDatasourceToolkit::Datasource.new
     else
       datasource = ForestAdminDatasourceRpc::Datasource.new(options, schema, schema_polling)
-
-      # Setup cleanup hooks for proper schema polling client shutdown
-      setup_cleanup_hooks(datasource)
-
       datasource
-    end
-  end
-
-  def self.setup_cleanup_hooks(datasource)
-    # Register cleanup handler for graceful shutdown
-    at_exit do
-      datasource.cleanup
-    rescue StandardError => e
-      # Silently ignore errors during exit cleanup to prevent test pollution
-      warn "[RPCDatasource] Error during at_exit cleanup: #{e.message}" if $VERBOSE
-    end
-
-    # Handle SIGINT (Ctrl+C)
-    Signal.trap('INT') do
-      begin
-        ForestAdminAgent::Facades::Container.logger&.log('Info', '[RPCDatasource] Received SIGINT, cleaning up...')
-      rescue StandardError
-        # Logger might not be available
-      end
-      datasource.cleanup
-      exit(0)
-    end
-
-    # Handle SIGTERM (default kill signal)
-    Signal.trap('TERM') do
-      begin
-        ForestAdminAgent::Facades::Container.logger&.log('Info', '[RPCDatasource] Received SIGTERM, cleaning up...')
-      rescue StandardError
-        # Logger might not be available
-      end
-      datasource.cleanup
-      exit(0)
     end
   end
 end